git clone should be git pull (read extended desc)
Git docs state that `git pull` should be used as an update tool and `git clone` only once to download from a remote, or if you want multiple working copies. Moreover, it should not be used to deploy to prod as an update tool as this uses up space on the end user's computer. The `master` branch should work off the spot, without needing to `git clone` other branches.