Cursed Civilian
I was quite unhappy about this role. The Cursed Civilian should change into a werewolf when they are attacked by werewolves, not anytime! This is THE example why we want the Game Masters to have a confirmation thing of the bot.

Also, as the on_death-function didn't have a return thing in it - would it actually change the player? Or would it make them dead, give 'em the dead participant role, THEN turn 'em back to a werewolf, giving them access to all cc's and allow them to talk with the dead people in graveyard?